Move _chunk.py into _feats.py
